Move generator handlers out of main
